Add files via upload
Farmer Assist Application is a online selling goods application based on blockchain technology. Blockchain is a system of recording information in a way that makes it difficult or impossible to change, hack, or cheat the system. This technology use to store the information about the farmers and market rates such as Farmer name, Farmer mobile no, Product name, Product cost, Product quantity etc.